requirements.txt
boto3==1.17.73
gunicorn==20.0.4
huey==2.3.0
itsdangerous==1.1.0
psycopg2-binary==2.8.5
python-dotenv==0.14.0
redis==3.5.3
requests==2.25.1
six==1.16.0
slackclient==1.3.2
urllib3==1.26.4
phonenumbers==8.12.9
Django==3.2.1
djangorestframework==3.11.1
django-extensions==3.0.8
django-environ==0.4.5
django-phone-verify==2.0.0
django-money==1.3.1
django-cors-headers==3.5.0
django-storages==1.10
djangorestframework-simplejwt==4.4.0
Pillow==7.2.0
sentry-sdk==1.0
stripe==2.51.0
django-hosts==4.0
whitenoise==5.2.0
onesignal-sdk==2.0.0
celery==5.0.5
python-telegram-bot==13.4.1
django-request-logging==0.7.2
messagebird==1.5.0
python-binance==1.0.7
stellar-sdk==3.3.3
django-cryptography==1.0
drf-nested-routers==0.93.3
django-silk==4.1.0
hashids==1.3.1
